Move Spaceport Animation code into its own file
