+++ src/game/CMakeLists.txt
@@ -87,6 +87,7 @@ set(game_sources
   roster_entry.cpp
   rush.cpp
   settings.cpp
+  spot.cpp
   start.cpp
   state_utils.cpp
   utils.cpp
